ps_backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
django-allauth==65.13.0
django-jazzmin==3.0.1
idna==3.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.5
urllib3==2.5.0
channels==4.1.0
daphne==4.1.2
djangorestframework-simplejwt==5.3.1
